# Manual parser for umask values.