--- a/src/utils/umask.rs
+++ b/src/utils/umask.rs
@@ -30,7 +30,6 @@ use std::{
 };
 
 use libc::mode_t;
-use regex::Regex;
 
 /// Get the current process's umask from `/proc/thread-self/status`.
 // There has been a Umask: field in /proc/self/status since Linux 4.7.
@@ -39,14 +38,6 @@ use regex::Regex;
 //NOTE: While the umask is shared between threads, it unshared with `CLONE_FS`
 //      so a single thread could have a different umask to other threads.
 fn get_umask_procfs(procfs: &ProcfsHandle) -> Result<Option<mode_t>, Error> {
-    // MSRV(1.70): Use OnceLock.
-    // MSRV(1.80): Use LazyLock.
-    // TODO: Figure out if we even need to use a regex for this. Surely there's
-    //       something like sscanf which works properly in Rust...
-    lazy_static! {
-        static ref RE: Regex = Regex::new(r"^Umask:\s*(0[0-7]+)$").unwrap();
-    }
-
     let status_file = procfs.open(ProcfsBase::ProcThreadSelf, "status", OpenFlags::O_RDONLY)?;
     let reader = BufReader::new(status_file);
     for line in reader.lines() {
@@ -54,11 +45,9 @@ fn get_umask_procfs(procfs: &ProcfsHandle) -> Result<Option<mode_t>, Error> {
             operation: "read lines from /proc/self/status".into(),
             source: err,
         })?;
-        // MSRV(1.65): Use let-else here.
-        if let Some((_, [umask])) = RE.captures(&line).map(|caps| caps.extract()) {
-            return Ok(Some(
-                mode_t::from_str_radix(umask, 8).expect("parsing 0[0-7]+ octal should work"),
-            ));
+
+        if let Some(umask) = line.strip_prefix("Umask:") {
+            return Ok(mode_t::from_str_radix(umask.trim(), 8).ok());
         }
     }
     Ok(None)