This patch changes the implementation of `get_umask_procfs` to avoid the dependency on the `regex` crate.

The original expression `^Umask:\s*(0[0-7]+)$` is implemented as:

```rust
if let Some(umask) = line.strip_prefix("Umask:") {
    return Ok(mode_t::from_str_radix(umask.trim(), 8).ok());
}
```

The previous code was using `expect` to get the result from `from_str_radix`. Now it just converts the `Result` from `from_str_radix` to an `Option`. The logic is different, but the behaviour is the same:

* If the value is correct (like `Umask: 0022`), both `Ok(Some(from_str_radix().expect()))` and  `Ok(from_str_radix().ok())` returns the same value.
* If the value is unexpected (like `Umask: 0999`),
    * The previous code will ignore the line (it does not match the regular expression), and will return `Ok(None)` after reading the `/proc/self/status` file.
    * The new code will try to parse `0999` as an octal. It will return an error, which is then converted to `None` by `ok()`.

## Build Improvements

After removing the `regex` crate, the size of the library and the compilation time improve significantly. 

|                     |  Before |     Now |
|---------------------|--------:|--------:|
| `libpathrs.so` size | `2.2 M` | `608 K` |
| Build time          |  `36 s` |  `16 s` |

* Before the patch:

    ```console
    $ hyperfine -w 1 -r 4 -p 'cargo clean || :' 'make release'
    Benchmark 1: make release
      Time (mean ± σ):     36.602 s ±  0.050 s    [User: 61.361 s, System: 4.345 s]
      Range (min … max):   36.552 s … 36.669 s    4 runs

    $ ls -sh target/release/libpathrs.so
    2.2M target/release/libpathrs.so
    ```

* After the patch:

    ```console
    $ hyperfine -w 1 -r 4 -p 'cargo clean || :' 'make release'
    Benchmark 1: make release
      Time (mean ± σ):     16.919 s ±  0.129 s    [User: 26.672 s, System: 2.974 s]
      Range (min … max):   16.744 s … 17.035 s    4 runs

    $ ls -sh target/release/libpathrs.so
    608K target/release/libpathrs.so
    ```