@cyphar 2024-09-25
This is nicer, though it seems we might remove this code entirely because it turns out checking the mode this strictly is kind of buggy.

See https://github.com/cyphar/filepath-securejoin/pull/29 and https://github.com/opencontainers/runc/issues/4408.