@ayosec 2024-09-25
> This is nicer, though it seems we might remove this code entirely because it turns out checking the mode this strictly is kind of buggy.

No problem!

Feel free to close the pull-request if you plan to remove the umask parser before the next release.